tests(mobx-react-lite): remove tests based on FinalizationRegistry
