+++ packages/mobx-react-lite/__tests__/strictAndConcurrentMode.test.tsx
@@ -7,6 +7,7 @@ import { useObserver } from "../src/useObserver"
 import { requestAnimationFrameMock } from "./utils/RequestAnimationFrameMockSession"
 
 afterEach(cleanup)
+afterEach(() => requestAnimationFrameMock.reset())
 
 test("uncommitted observing components should not attempt state changes", () => {
     const store = mobx.observable({ count: 0 })
@@ -115,4 +116,40 @@ test("destroy reaction in the next animation frame if component destroyed", asyn
     expect(unobserved).toBeCalledTimes(1)
 
     doneCallback()
+})
+
+test("uncommitted components should not leak observations", async () => {
+    const store = mobx.observable({ count1: 0, count2: 0 })
+
+    // Track whether counts are observed
+    let count1IsObserved = false
+    let count2IsObserved = false
+    mobx.onBecomeObserved(store, "count1", () => (count1IsObserved = true))
+    mobx.onBecomeUnobserved(store, "count1", () => (count1IsObserved = false))
+    mobx.onBecomeObserved(store, "count2", () => (count2IsObserved = true))
+    mobx.onBecomeUnobserved(store, "count2", () => (count2IsObserved = false))
+
+    const TestComponent1 = () => useObserver(() => <div>{store.count1}</div>)
+    const TestComponent2 = () => useObserver(() => <div>{store.count2}</div>)
+
+    // Render, then remove only #2
+    const rendering = render(
+        <React.StrictMode>
+            <TestComponent1 />
+            <TestComponent2 />
+        </React.StrictMode>
+    )
+    rendering.rerender(
+        <React.StrictMode>
+            <TestComponent1 />
+        </React.StrictMode>
+    )
+
+    // Force reactions to be disposed
+    requestAnimationFrameMock.triggerAllAnimationFrames()
+
+    // count1 should still be being observed by Component1,
+    // but count2 should have had its reaction cleaned up.
+    expect(count1IsObserved).toBeTruthy()
+    expect(count2IsObserved).toBeFalsy()
 })
